move connect to postgres package
